functions/ai_processor/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
google-genai==1.1.0 